Implement basic functionality to add (dummy) labels to pull requests
For now, the action's labeling logic is very silly: it adds an `even`
label or an `odd` label according to the PR number. We'll ultimately
replace this with the real logic, but this is a sufficient proof of
concept to demonstrate that the process of adding labels is working.

The `getPrNumber` and `addLabels` functions were copied from
https://github.com/actions/labeler, an action created by GitHub for
labeling pull requests based on their changed files.