Validate teams configuration YAML & parse it into a label to members map
This ensures the YAML we loaded has the expected structure, which is an
object mapping team names to team metadata. The metadata must contain a
`members` array of objects which must include a `github` property whose
value is a GitHub username. The metadata can also include a `short_name`
string, which will ultimately used as the PR label if it exists
(otherwise the team name will be used). Any other fields may be present,
but they'll be ignored by this action.